Update build to work with OpenSSL 3.2 (#642)
From recent builds, it appears that RHEL 9 has upgraded from OpenSSL 3.0 to 3.2, which breaks our builds. I found two places where we filter on the OpenSSL version.

In Makefile, I removed the filter because we no longer support OpenSSL 1.0, which is why the case statement was added. All versions of OpenSSL we support should behave the same, so there's no need to filter on the version.

In contrib/debian/rules, I relaxed the filter to handle any 3.x version of OpenSSL.