Avoid container build by adding /.mariner-toolkit-ignore-dockerenv (#646)
Our Azure Linux builds started failing in the past week or so because Azure Linux recently made a change to how they detect container builds (see https://github.com/microsoft/azurelinux/pull/11135), so the trick we employed to force a regular build (even though we're building in a container) stopped working. This change removes the trick and takes the recommended approach of adding `/.mariner-toolkit-ignore-dockerenv` to the container.