Fix sed -i commands in build script (#649)
Some sed commands in one of the build scripts combine the -i (edit files in place) and -e (expression to be executed) options as -ie. This changes the behavior of -i to make a backup of the edited file with 'e' as the suffix. Because these sed commands only had one expression statement on the command line, the expression was still evaluated even though the -e argument was not present. So the result was effectively the same, which the side effect of creating a backup file that no one asked for.

This change updates the sed commands just to avoid confusion.