Add support for handling network errors from parent edgeHub (#651)
In nested mode, child devices are not able to detect if the network between the parent and IoT Hub is down. This leads to some incorrect behavior where the child assumes IoT Hub is available during its startup. edgeHub will return a 500 Internal Server Error in this case. This PR adds handling of 500 Internal Server Errors from edgeHub as network errors so identity service starts correctly when connection to IoT Hub is down.

Also improves the logging of HTTP errors by logging the HTTP error code instead of a JSON parse error.