fix: workaround for snapcraft needing a user runtime directory (#660)
This comes from discussion with Michael Croft-White on Matrix: https://matrix.to/#/!GGqzbFAUQprdPgYYCM:ubuntu.com/$6hguIiRWB8jEq-J7JgZDamYOaFIph9mZpK-Nz8uWoRY

While this isn't a long-term fix, this should provide a workaround to allow the snap to build.

Also includes an unrelated change to the Azure Linux builds, which was needed to pass the PR checks because another build failure made its way into the codebase while the builds were broken.