Make minor updates to the packages workflow (#659)
The purpose of this change was originally to update the go dependency in our Azure Linux builds, but that fix went into d1400a7270bb68cd4cae1fcb3963f6564106071a to unblock the build. What's left are a few minor changes that needed to be made, specifically using more appropriate runner images and an updated GitHub Action in the packages workflow.